+++ api/batch_interpolate.sh
@@ -76,6 +76,7 @@ else
 	if [[ $COUNT -gt 0 ]] ; then
 		VIDEOFILES=`find input/vr/interpolate -maxdepth 1 -type f -name '*.mp4' -o -name '*.webm'`
 		for nextinputfile in $VIDEOFILES ; do
+			[ -e user/default/comfyui_stereoscopic/.pipelinepause ] && exit 0
 			INDEX+=1
 			echo "$INDEX/$COUNT" >input/vr/interpolate/BATCHPROGRESS.TXT
 			echo "interpolate" >user/default/comfyui_stereoscopic/.daemonstatus
