+++ api/batch_singleloop.sh
@@ -59,6 +59,7 @@ else
 	if [[ $COUNT -gt 0 ]] ; then
 		VIDFILES=`find input/vr/singleloop -maxdepth 1 -type f -name '*.mp4' -o -name '*.webm'`
 		for nextinputfile in $VIDFILES ; do
+			[ -e user/default/comfyui_stereoscopic/.pipelinepause ] && exit 0
 			INDEX+=1
 			echo "$INDEX/$COUNT" >input/vr/singleloop/BATCHPROGRESS.TXT
 			echo "singleloop" >user/default/comfyui_stereoscopic/.daemonstatus
